// frontend/src/core/api/generated/model/taskViewPriority.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type TaskViewPriority = (typeof TaskViewPriority)[keyof typeof TaskViewPriority]

export const TaskViewPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
} as const