// frontend/src/core/api/generated/model/taskViewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type TaskViewStatus = (typeof TaskViewStatus)[keyof typeof TaskViewStatus]

export const TaskViewStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
} as const